.3669509757995524:307ce2b82abd8b9472a703046aea43a5_69ef14b5596d393c3aa1a888.69f5e3f4363b740c909a29e9.69f5e3f499e735f5364c707b:Trae CN.T(2026/5/2 19:45:56)

+++ include/combinetool/io/stream_writer.h
@@ -13,7 +13,7 @@ namespace io {
 
 struct WriteResult {
     bool success;
-    size_t bytesWritten;
+    uint64_t bytesWritten;
     std::string errorMessage;
 };
